# docs/swagger.yaml
definitions:
...
  models.Order:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      ownerID:
        type: string
      total:
        type: number
    type: object
  models.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/orders:
    get:
      description: Возвращает список всех заказов (только для админа)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить все заказы (админ)
      tags:
      - Админ / Заказы
  /api/admin/orders/export:
    get:
      description: Экспортирует все заказы в формате CSV (только для админа)
      produces:
      - text/csv
      responses:
        "200":
          description: CSV файл
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Экспорт заказов в CSV (админ)
      tags:
      - Админ / Заказы
  /api/admin/products:
    post:
      consumes:
      - application/json
      description: Создаёт новый товар (только для администратора)
      parameters:
      - description: Данные товара
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Добавить товар (админ)
      tags:
      - Товар
  /api/admin/products/{id}:
    delete:
      description: Удаляет товар по его ID (только для администратора)
      parameters:
      - description: ID товара
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Product deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить товар (админ)
      tags:
      - Товар
    put:
      consumes:
      - application/json
      description: Обновляет существующий товар по его ID (только для администратора)
      parameters:
      - description: ID товара
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные товара
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - text/plain
      responses:
        "200":
          description: Product updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить товар (админ)
      tags:
      - Товар
...
  /api/order:
    post:
      description: Оформляет заказ из текущей корзины owner_id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Оформить заказ
      tags:
      - Заказ
  /api/orders:
    get:
      description: Возвращает список заказов для текущего owner_id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Получить заказы пользователя
      tags:
      - Заказ
  /api/products:
    get:
      description: Получает список товаров с возможностью фильтрации
      parameters:
      - description: Поиск по названию
        in: query
        name: search
        type: string
      - description: Фильтр по категории
        in: query
        name: category
        type: string
      - description: Минимальная цена
        in: query
        name: min_price
        type: number
      - description: Максимальная цена
        in: query
        name: max_price
        type: number
      - description: Сортировка
        in: query
        name: sort
        type: string
      - description: Ограничение количества результатов
        in: query
        name: limit
        type: integer
      - description: Сдвиг для пагинации
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Получить список товаров
      tags:
      - Товар
  /api/products/{id}:
    get:
      description: Возвращает детали товара по его идентификатору
      parameters:
      - description: ID товара
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Получить товар по ID
      tags:
      - Товар
...
swagger: "2.0"